tests/test_gRPC_service/service_tuning/requirements.txt
fastmcp
grpcio
grpcio-tools
protobuf
llama-index
langchain
langchain-community
transformers
torch
datasets
peft
accelerate
bitsandbytes
trl
numpy
pandas
scikit-learn
pydantic
python-dotenv
asyncio
uvloop
